example.py: set verify_ssl=False, since SMA Inverters use invalid certs
SMA Inverters use self-signed certificates which are untrusted by default on Python.

Without this change, attempting to use example.py with HTTPS fails with:

  $ python example.py https://sma.local user 1234
  DEBUG:asyncio:Using selector: EpollSelector
  ERROR:pysma:Could not start session, Could not connect to SMA at https://sma.local (timeout), got {}
  INFO:__main__:No session ID

With this change, connecting works as expected.

In an ideal world, using a trusted TLS certificate would be the best fix. However, I'm unaware of a way to upload a trusted TLS cert to an SMA inverter. As such, setting verify_ssl=False seems to be the only practical way to use HTTPS with this hardware.